+++ package.json
@@ -15,6 +15,7 @@
     "flatpickr": "4.5.7",
     "lodash": "4.17.11",
     "moment": "2.24.0",
+    "style-loader": "^0.23.1",
     "webpack": "4.29.5",
     "webpack-cli": "3.2.3",
     "webpack-dev-server": "3.2.0"
