+++ public/index.html
@@ -8,7 +8,6 @@
   <link href="https://fonts.googleapis.com/css?family=Montserrat:500,600,700,800,900&amp;subset=cyrillic-ext" rel="stylesheet">
   <link rel="stylesheet" type="text/css" media="screen" href="./css/normalize.css" />
   <link rel="stylesheet" type="text/css" media="screen" href="./css/main.css" />
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/flatpickr/dist/flatpickr.min.css">
 </head>
 <body>
 
@@ -253,3 +252,4 @@
 <script src="./bundle.js"></script>
 </body>
 </html>
+
