+++ src/lib/constans.js
@@ -1,0 +1,16 @@
+export const FILTER_PROPS = [{
+  value: `everything`,
+  checked: `checked`,
+  disabled: ``
+},
+{
+  value: `future`,
+  checked: ``,
+  disabled: ``
+},
+{
+  value: `past`,
+  checked: ``,
+  disabled: ``
+}
+];
