+++ webpack.config.js
@@ -7,6 +7,20 @@ module.exports = {
     filename: `bundle.js`,
     path: path.join(__dirname, `public`)
   },
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        use: [
+          {
+            loader: `style-loader`
+          },
+          {
+            loader: `css-loader`
+          }
+        ]}
+    ]
+  },
   devtool: `source-map`,
   devServer: {
     contentBase: path.join(__dirname, `public`),
